Move test helper class into module
